pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "ai-core-sdk"
version = "0.1.0"
description = "Enterprise Agentic AI Platform (EAAP) core SDK — DI-driven, async, observable, policy-aware."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "EAAP Platform Team" }]
keywords = ["agentic", "ai", "langgraph", "mcp", "litellm", "enterprise", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    # --- DI & validation ---
    "injector>=0.21.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # --- Async runtime / HTTP ---
    "anyio>=4.3.0",
    "httpx>=0.27.0",

    # --- Persistence (SQL) ---
    "SQLAlchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # --- LLM proxy & resilience ---
    "litellm>=1.35.0",
    "tenacity>=8.2.3",

    # --- Orchestration ---
    "langgraph>=0.0.55",
    "fastmcp>=0.2.0",

    # --- API / Security ---
    "fastapi>=0.110.0",
    "starlette>=0.36.0",
    "PyJWT[crypto]>=2.8.0",

    # --- Observability ---
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "langfuse>=2.20.0",

    # --- CLI scaffolding ---
    "typer>=0.12.0",
    "rich>=13.7.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
storage-aws = [
    "aioboto3>=12.3.0",
    "boto3>=1.34.0",
]
secrets-aws = [
    "aioboto3>=12.3.0",
]
vector-pgvector = [
    "pgvector>=0.2.5",
]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML",
    "respx>=0.21.0",
]

[project.scripts]
eaap = "ai_core.cli.main:app"

[project.urls]
Homepage = "https://github.com/eaap/ai-core-sdk"
Issues = "https://github.com/eaap/ai-core-sdk/issues"

# --------------------------------------------------------------------------
# Build / packaging
# --------------------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/ai_core"]

[tool.hatch.build.targets.wheel.force-include]
"src/ai_core/cli/templates" = "ai_core/cli/templates"

[tool.hatch.build.targets.sdist]
include = [
    "src/ai_core",
    "tests",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

# --------------------------------------------------------------------------
# Tooling
# --------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: fast, isolated unit tests",
    "component: in-process tests across multiple SDK modules",
    "integration: tests that hit real external systems (DB, OPA, OTel, LLM)",
]
filterwarnings = [
    # Short HMAC secrets are fine in unit tests; we test correctness, not crypto strength.
    "ignore::jwt.warnings.InsecureKeyLengthWarning",
]

[tool.coverage.run]
branch = true
source = ["src/ai_core"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle / pyflakes
    "I",                 # isort
    "N",                 # pep8-naming
    "UP",                # pyupgrade
    "B",                 # bugbear
    "ANN",               # annotations
    "ASYNC",             # async correctness
    "S",                 # bandit (security)
    "C4",                # comprehensions
    "RET",               # return statements
    "SIM",               # simplifications
    "TCH",               # type-checking
    "PL",                # pylint
    "RUF",               # ruff-specific
]
ignore = [
    "ANN101", "ANN102",  # self/cls annotations
    "PLR0913",           # too many arguments (DI constructors are wide)
    "S101",              # asserts (used in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S", "PLR2004"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["fastmcp.*", "langfuse.*", "litellm.*", "injector.*"]
ignore_missing_imports = true
